# perf: add comprehensive BenchmarkDotNet performance suite with CI integration