--- a/benchmarks/FastGeoMesh.Benchmarks/BenchmarkDotNet.Artifacts/results/FastGeoMesh.Benchmarks.RectangleVsTessellationBenchmarks-report.html
+++ b/benchmarks/FastGeoMesh.Benchmarks/BenchmarkDotNet.Artifacts/results/FastGeoMesh.Benchmarks.RectangleVsTessellationBenchmarks-report.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang='en'>
+<head>
+<meta charset='utf-8' />
+<title>FastGeoMesh.Benchmarks.RectangleVsTessellationBenchmarks-20250928-223137</title>
+
+<style type="text/css">
+	table { border-collapse: collapse; display: block; width: 100%; overflow: auto; }
+	td, th { padding: 6px 13px; border: 1px solid #ddd; text-align: right; }
+	tr { background-color: #fff; border-top: 1px solid #ccc; }
+	tr:nth-child(even) { background: #f8f8f8; }
+</style>
+</head>
+<body>
+<pre><code>
+BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.6584)
+Intel Core Ultra 7 165U, 1 CPU, 14 logical and 12 physical cores
+.NET SDK 9.0.304
+  [Host]     : .NET 8.0.20 (8.0.2025.41914), X64 RyuJIT AVX2
+  DefaultJob : .NET 8.0.20 (8.0.2025.41914), X64 RyuJIT AVX2
+</code></pre>
+<pre><code></code></pre>
+
+<table>
+<thead><tr><th>Method       </th><th>Mean</th><th>Error</th><th>StdDev</th><th>Median</th><th>Ratio</th><th>RatioSD</th><th>Gen0</th><th>Gen1</th><th>Allocated</th><th>Alloc Ratio</th>
+</tr>
+</thead><tbody><tr><td>RectangleFastPath</td><td>28.98 &mu;s</td><td>7.276 &mu;s</td><td>21.454 &mu;s</td><td>14.52 &mu;s</td><td>1.60</td><td>1.58</td><td>17.9596</td><td>5.3864</td><td>110.11 KB</td><td>1.00</td>
+</tr><tr><td>HexagonTessellation</td><td>18.96 &mu;s</td><td>1.468 &mu;s</td><td>4.328 &mu;s</td><td>21.01 &mu;s</td><td>1.05</td><td>0.61</td><td>9.6130</td><td>1.3428</td><td>59.06 KB</td><td>0.54</td>
+</tr></tbody></table>
+</body>
+</html>