--- a/benchmarks/FastGeoMesh.Benchmarks/FastGeoMesh.Benchmarks.csproj
+++ b/benchmarks/FastGeoMesh.Benchmarks/FastGeoMesh.Benchmarks.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <WarningsAsErrors>IDE0160;IDE0052;IDE0054</WarningsAsErrors>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="../../src/FastGeoMesh/FastGeoMesh.csproj" />
+  </ItemGroup>
+</Project>