## Performance Benchmarking Infrastructure
Adds comprehensive performance benchmarking suite using BenchmarkDotNet to establish performance baselines and identify optimization opportunities.

### What's Added
#### Benchmark Suite (`benchmarks/FastGeoMesh.Benchmarks/`) 
- **Rectangle vs Tessellation**: Compares axis-aligned rectangle fast-path vs generic polygon tessellation 
- **Mesh Size Scaling**: Performance impact of geometry size and target edge length (parameterized) 
- **Quad Quality Scoring**: Impact of quality thresholds on tessellation and triangulation fallback 
- **IndexedMesh Operations**: Mesh conversion and adjacency building performance
- 
#### Key Features 
- **Memory diagnostics**: Tracks GC pressure and allocation patterns 
- **Parameterized benchmarks**: Multiple geometry sizes and quality thresholds 
- **CLI interface**: `dotnet run -c Release -- [benchmark-name]` - **CI integration**: Automatic benchmarks on main branch pushes with artifact storage
- 
### Initial Performance Insights
#### Surprising Rectangle Results 
- Rectangle "fast-path": **28.98 μs** (110.11 KB allocated) 
- Hexagon tessellation: **18.96 μs** (59.06 KB allocated) 
- **Finding**: Rectangle path is 1.5x slower than generic tessellation

#### Scaling Characteristics 
- **Super-linear scaling**: TargetEdgeLength reduction 4x = 40x+ execution time 
- **Memory pressure**: 100MB+ allocations for large meshes trigger Gen2 GC 
- **Sweet spot**: TargetEdge=1.0, Size≤50 for interactive performance
- 
#### Performance Matrix 
``` TargetEdgeLength=1.0:   Size=10   →   23.60 μs    (62.79 KB) Size=50   →  2,608 μs    (1.57 MB)
Size=100  →  5,038 μs    (6.26 MB)
TargetEdgeLength=0.25:  Size=100  → 193,844 μs  (100 MB) ```

### Next Steps Identified 
1. **Investigate rectangle fast-path performance** - why slower than tessellation? 
3. **Memory optimization opportunities** - reduce allocations for large meshes 
4. 3. **Algorithmic improvements** - address super-linear scaling
5. 
### Technical Details 
- **BenchmarkDotNet 0.14.0** with .NET 8 optimization 
- **MemoryDiagnoser** enabled for allocation tracking 
- **Baseline comparisons** for relative performance measurement 
- **Statistical validation** with confidence intervals and outlier detection
- 
This establishes the foundation for data-driven performance optimization and regression detection.